requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
msal==1.31.0
requests==2.32.3
openai==1.47.0
supabase==2.6.0
pydantic==2.8.2
beautifulsoup4==4.12.3
tiktoken==0.7.0
pypdf==4.3.1
python-multipart==0.0.9
tenacity==9.0.0
orjson==3.10.7

python-docx==1.1.2
